+++ include/ui/newmappopup.h
@@ -22,10 +22,10 @@ public:
     bool existingLayout;
     bool importedMap;
     QString layoutId;
-    void init(int, int, QString, QString);
-    void initImportMap(MapLayout *);
-    void useLayout(QString);
-    void connectSignals();
+    void init();
+    void init(MapSortOrder type, QVariant data);
+    void init(MapLayout *);
+    static void setDefaultSettings(Project *project);
 
 signals:
     void applied();
@@ -33,11 +33,31 @@ signals:
 private:
     Ui::NewMapPopup *ui;
     Project *project;
-    void setDefaultValues(int, QString);
-    void setDefaultValuesImportMap(MapLayout *);
-    void setDefaultValuesProjectConfig(bool, MapLayout*);
     bool checkNewMapDimensions();
     bool checkNewMapGroup();
+    void saveSettings();
+    void useLayout(QString layoutId);
+    void useLayoutSettings(MapLayout *mapLayout);
+
+    struct Settings {
+        QString group;
+        int width;
+        int height;
+        int borderWidth;
+        int borderHeight;
+        QString primaryTilesetLabel;
+        QString secondaryTilesetLabel;
+        QString type;
+        QString location;
+        QString song;
+        bool canFlyTo;
+        bool showLocationName;
+        bool allowRunning;
+        bool allowBiking;
+        bool allowEscaping;
+        int floorNumber;
+    };
+    static struct Settings settings;
 
 private slots:
     void on_pushButton_NewMap_Accept_clicked();