+++ src/ui/newmappopup.cpp
@@ -9,6 +9,8 @@
 #include <QPalette>
 #include <QStringList>
 
+struct NewMapPopup::Settings NewMapPopup::settings = {};
+
 NewMapPopup::NewMapPopup(QWidget *parent, Project *project) :
     QMainWindow(parent),
     ui(new Ui::NewMapPopup)
@@ -21,30 +23,104 @@ NewMapPopup::NewMapPopup(QWidget *parent, Project *project) :
 
 NewMapPopup::~NewMapPopup()
 {
+    saveSettings();
     delete ui;
 }
 
-void NewMapPopup::init(int type, int group, QString sec, QString layoutId) {
-    switch (type)
-    {
-        case MapSortOrder::Group:
-            setDefaultValues(group, QString());
-            break;
-        case MapSortOrder::Area:
-            setDefaultValues(group, sec);
-            break;
-        case MapSortOrder::Layout:
-            useLayout(layoutId);
-            setDefaultValues(group, QString());
-            break;
-    }
-    connectSignals();
+void NewMapPopup::init() {
+    // Populate combo boxes
+    ui->comboBox_NewMap_Primary_Tileset->addItems(project->primaryTilesetLabels);
+    ui->comboBox_NewMap_Secondary_Tileset->addItems(project->secondaryTilesetLabels);
+    ui->comboBox_NewMap_Group->addItems(project->groupNames);
+    ui->comboBox_NewMap_Song->addItems(project->songNames);
+    ui->comboBox_NewMap_Type->addItems(project->mapTypes);
+    ui->comboBox_NewMap_Location->addItems(project->mapSectionValueToName.values());
+
+    // Set spin box limits
+    ui->spinBox_NewMap_Width->setMinimum(1);
+    ui->spinBox_NewMap_Height->setMinimum(1);
+    ui->spinBox_NewMap_Width->setMaximum(project->getMaxMapWidth());
+    ui->spinBox_NewMap_Height->setMaximum(project->getMaxMapHeight());
+    ui->spinBox_NewMap_BorderWidth->setMinimum(1);
+    ui->spinBox_NewMap_BorderHeight->setMinimum(1);
+    ui->spinBox_NewMap_Floor_Number->setMinimum(-128);
+    ui->spinBox_NewMap_Floor_Number->setMaximum(127);
+
+    // Hide config specific ui elements
+    bool hasFlags = (projectConfig.getBaseGameVersion() != BaseGameVersion::pokeruby);
+    ui->checkBox_NewMap_Allow_Running->setVisible(hasFlags);
+    ui->checkBox_NewMap_Allow_Biking->setVisible(hasFlags);
+    ui->checkBox_NewMap_Allow_Escape_Rope->setVisible(hasFlags);
+    ui->label_NewMap_Allow_Running->setVisible(hasFlags);
+    ui->label_NewMap_Allow_Biking->setVisible(hasFlags);
+    ui->label_NewMap_Allow_Escape_Rope->setVisible(hasFlags);
+
+    bool hasCustomBorders = projectConfig.getUseCustomBorderSize();
+    ui->spinBox_NewMap_BorderWidth->setVisible(hasCustomBorders);
+    ui->spinBox_NewMap_BorderHeight->setVisible(hasCustomBorders);
+    ui->label_NewMap_BorderWidth->setVisible(hasCustomBorders);
+    ui->label_NewMap_BorderHeight->setVisible(hasCustomBorders);
+
+    bool hasFloorNumber = projectConfig.getFloorNumberEnabled();
+    ui->spinBox_NewMap_Floor_Number->setVisible(hasFloorNumber);
+    ui->label_NewMap_Floor_Number->setVisible(hasFloorNumber);
+
+    // Restore previous settings
+    ui->lineEdit_NewMap_Name->setText(project->getNewMapName());
+    ui->comboBox_NewMap_Group->setTextItem(settings.group);
+    ui->spinBox_NewMap_Width->setValue(settings.width);
+    ui->spinBox_NewMap_Height->setValue(settings.height);
+    ui->spinBox_NewMap_BorderWidth->setValue(settings.borderWidth);
+    ui->spinBox_NewMap_BorderHeight->setValue(settings.borderHeight);
+    ui->comboBox_NewMap_Primary_Tileset->setTextItem(settings.primaryTilesetLabel);
+    ui->comboBox_NewMap_Secondary_Tileset->setTextItem(settings.secondaryTilesetLabel);
+    ui->comboBox_NewMap_Type->setTextItem(settings.type);
+    ui->comboBox_NewMap_Location->setTextItem(settings.location);
+    ui->comboBox_NewMap_Song->setTextItem(settings.song);
+    ui->checkBox_NewMap_Flyable->setChecked(settings.canFlyTo);
+    ui->checkBox_NewMap_Show_Location->setChecked(settings.showLocationName);
+    ui->checkBox_NewMap_Allow_Running->setChecked(settings.allowRunning);
+    ui->checkBox_NewMap_Allow_Biking->setChecked(settings.allowBiking);
+    ui->checkBox_NewMap_Allow_Escape_Rope->setChecked(settings.allowEscaping);
+    ui->spinBox_NewMap_Floor_Number->setValue(settings.floorNumber);
+
+    // Connect signals
+    connect(ui->spinBox_NewMap_Width, QOverload<int>::of(&QSpinBox::valueChanged), [=](int){checkNewMapDimensions();});
+    connect(ui->spinBox_NewMap_Height, QOverload<int>::of(&QSpinBox::valueChanged), [=](int){checkNewMapDimensions();});
+
+    ui->frame_NewMap_Options->setEnabled(true);
 }
 
-void NewMapPopup::initImportMap(MapLayout *mapLayout) {
+// Creating new map by right-clicking in the map list
+void NewMapPopup::init(MapSortOrder type, QVariant data) {
+    switch (type)
+    {
+    case MapSortOrder::Group:
+        settings.group = project->groupNames.at(data.toInt());
+        break;
+    case MapSortOrder::Area:
+        settings.location = data.toString();
+        break;
+    case MapSortOrder::Layout:
+        useLayout(data.toString());
+        break;
+    }
+    init();
+}
+
+// Creating new map from AdvanceMap import
+void NewMapPopup::init(MapLayout *mapLayout) {
     this->importedMap = true;
-    setDefaultValuesImportMap(mapLayout);
-    connectSignals();
+    useLayoutSettings(mapLayout);
+
+    this->map = new Map();
+    this->map->layout = new MapLayout();
+    this->map->layout->blockdata = mapLayout->blockdata;
+
+    if (!mapLayout->border.isEmpty()) {
+        this->map->layout->border = mapLayout->border;
+    }
+    init();
 }
 
 bool NewMapPopup::checkNewMapDimensions() {
@@ -53,7 +129,6 @@ bool NewMapPopup::checkNewMapDimensions() {
 
     if (numMetatiles > maxMetatiles) {
         ui->frame_NewMap_Warning->setVisible(true);
-        //ui->label_NewMap_WarningMessage->setText("WARNING: The specified map dimensions are too large.");
         QString errorText = QString("Error: The specified width and height are too large.\n"
                     "The maximum map width and height is the following: (width + 15) * (height + 14) <= %1\n"
                     "The specified map width and height was: (%2 + 15) * (%3 + 14) = %4")
@@ -89,148 +164,66 @@ bool NewMapPopup::checkNewMapGroup() {
     }
 }
 
-void NewMapPopup::connectSignals() {
-    ui->spinBox_NewMap_Width->setMinimum(1);
-    ui->spinBox_NewMap_Height->setMinimum(1);
-    ui->spinBox_NewMap_Width->setMaximum(project->getMaxMapWidth());
-    ui->spinBox_NewMap_Height->setMaximum(project->getMaxMapHeight());
+void NewMapPopup::setDefaultSettings(Project *project) {
+    settings.group = project->groupNames.at(0);
+    settings.width = project->getDefaultMapSize();
+    settings.height = project->getDefaultMapSize();
+    settings.borderWidth = DEFAULT_BORDER_WIDTH;
+    settings.borderHeight = DEFAULT_BORDER_HEIGHT;
+    settings.primaryTilesetLabel = project->getDefaultPrimaryTilesetLabel();
+    settings.secondaryTilesetLabel = project->getDefaultSecondaryTilesetLabel();
+    settings.type = project->mapTypes.at(0);
+    settings.location = project->mapSectionValueToName.values().at(0);
+    settings.song = project->songNames.at(0);
+    settings.canFlyTo = false;
+    settings.showLocationName = true;
+    settings.allowRunning = false;
+    settings.allowBiking = false;
+    settings.allowEscaping = false;
+    settings.floorNumber = 0;
+}
 
-    //ui->icon_NewMap_WarningIcon->setPixmap();
-    connect(ui->spinBox_NewMap_Width, QOverload<int>::of(&QSpinBox::valueChanged), [=](int){checkNewMapDimensions();});
-    connect(ui->spinBox_NewMap_Height, QOverload<int>::of(&QSpinBox::valueChanged), [=](int){checkNewMapDimensions();});
+void NewMapPopup::saveSettings() {
+    settings.group = ui->comboBox_NewMap_Group->currentText();
+    settings.width = ui->spinBox_NewMap_Width->value();
+    settings.height = ui->spinBox_NewMap_Height->value();
+    settings.borderWidth = ui->spinBox_NewMap_BorderWidth->value();
+    settings.borderHeight = ui->spinBox_NewMap_BorderHeight->value();
+    settings.primaryTilesetLabel = ui->comboBox_NewMap_Primary_Tileset->currentText();
+    settings.secondaryTilesetLabel = ui->comboBox_NewMap_Secondary_Tileset->currentText();
+    settings.type = ui->comboBox_NewMap_Type->currentText();
+    settings.location = ui->comboBox_NewMap_Location->currentText();
+    settings.song = ui->comboBox_NewMap_Song->currentText();
+    settings.canFlyTo = ui->checkBox_NewMap_Flyable->isChecked();
+    settings.showLocationName = ui->checkBox_NewMap_Show_Location->isChecked();
+    settings.allowRunning = ui->checkBox_NewMap_Allow_Running->isChecked();
+    settings.allowBiking = ui->checkBox_NewMap_Allow_Biking->isChecked();
+    settings.allowEscaping = ui->checkBox_NewMap_Allow_Escape_Rope->isChecked();
+    settings.floorNumber = ui->spinBox_NewMap_Floor_Number->value();
+}
+
+void NewMapPopup::useLayoutSettings(MapLayout *layout) {
+    if (!layout) return;
+    settings.width = layout->width;
+    settings.height = layout->height;
+    settings.borderWidth = layout->border_width;
+    settings.borderHeight = layout->border_height;
+    settings.primaryTilesetLabel = layout->tileset_primary_label;
+    settings.secondaryTilesetLabel = layout->tileset_secondary_label;
 }
 
 void NewMapPopup::useLayout(QString layoutId) {
     this->existingLayout = true;
     this->layoutId = layoutId;
-}
+    useLayoutSettings(project->mapLayouts.value(this->layoutId));
 
-void NewMapPopup::setDefaultValues(int groupNum, QString mapSec) {
-    ui->lineEdit_NewMap_Name->setText(project->getNewMapName());
-
-    ui->comboBox_NewMap_Primary_Tileset->addItems(project->tilesetLabels.value("primary"));
-    ui->comboBox_NewMap_Secondary_Tileset->addItems(project->tilesetLabels.value("secondary"));
-
-    ui->comboBox_NewMap_Group->addItems(project->groupNames);
-    ui->comboBox_NewMap_Group->setCurrentText(project->groupNames.at(groupNum));
-
-    ui->comboBox_Song->addItems(project->songNames);
-
-    if (existingLayout) {
-        ui->spinBox_NewMap_Width->setValue(project->mapLayouts.value(layoutId)->width);
-        ui->spinBox_NewMap_Height->setValue(project->mapLayouts.value(layoutId)->height);
-        ui->comboBox_NewMap_Primary_Tileset->setCurrentText(project->mapLayouts.value(layoutId)->tileset_primary_label);
-        ui->comboBox_NewMap_Secondary_Tileset->setCurrentText(project->mapLayouts.value(layoutId)->tileset_secondary_label);
-        ui->spinBox_NewMap_Width->setDisabled(true);
-        ui->spinBox_NewMap_Height->setDisabled(true);
-        ui->spinBox_NewMap_BorderWidth->setDisabled(true);
-        ui->spinBox_NewMap_BorderHeight->setDisabled(true);
-        ui->comboBox_NewMap_Primary_Tileset->setDisabled(true);
-        ui->comboBox_NewMap_Secondary_Tileset->setDisabled(true);
-    } else {
-        ui->spinBox_NewMap_Width->setValue(project->getDefaultMapSize());
-        ui->spinBox_NewMap_Height->setValue(project->getDefaultMapSize());
-        ui->spinBox_NewMap_BorderWidth->setValue(DEFAULT_BORDER_WIDTH);
-        ui->spinBox_NewMap_BorderHeight->setValue(DEFAULT_BORDER_HEIGHT);
-    }
-
-    ui->comboBox_NewMap_Type->addItems(project->mapTypes);
-    ui->comboBox_NewMap_Location->addItems(project->mapSectionValueToName.values());
-    if (!mapSec.isEmpty()) ui->comboBox_NewMap_Location->setCurrentText(mapSec);
-    ui->checkBox_NewMap_Show_Location->setChecked(true);
-
-    ui->frame_NewMap_Options->setEnabled(true);
-
-    setDefaultValuesProjectConfig(false, NULL);
-}
-
-void NewMapPopup::setDefaultValuesImportMap(MapLayout *mapLayout) {
-    ui->lineEdit_NewMap_Name->setText(project->getNewMapName());
-
-    ui->comboBox_NewMap_Primary_Tileset->addItems(project->tilesetLabels.value("primary"));
-    ui->comboBox_NewMap_Secondary_Tileset->addItems(project->tilesetLabels.value("secondary"));
-
-    ui->comboBox_NewMap_Group->addItems(project->groupNames);
-    ui->comboBox_NewMap_Group->setCurrentText(project->groupNames.at(0));
-
-    ui->comboBox_Song->addItems(project->songNames);
-
-    ui->spinBox_NewMap_Width->setValue(mapLayout->width);
-    ui->spinBox_NewMap_Height->setValue(mapLayout->height);
-    ui->comboBox_NewMap_Primary_Tileset->setCurrentText(mapLayout->tileset_primary_label);
-    ui->comboBox_NewMap_Secondary_Tileset->setCurrentText(mapLayout->tileset_secondary_label);
-
-    ui->comboBox_NewMap_Type->addItems(project->mapTypes);
-    ui->comboBox_NewMap_Location->addItems(project->mapSectionValueToName.values());
-    ui->checkBox_NewMap_Show_Location->setChecked(true);
-
-    ui->frame_NewMap_Options->setEnabled(true);
-
-    setDefaultValuesProjectConfig(true, mapLayout);
-
-    map = new Map();
-    map->layout = new MapLayout();
-    map->layout->blockdata = mapLayout->blockdata;
-
-    if (!mapLayout->border.isEmpty()) {
-        map->layout->border = mapLayout->border;
-    }
-}
-
-void NewMapPopup::setDefaultValuesProjectConfig(bool importedMap, MapLayout *mapLayout) {
-    switch (projectConfig.getBaseGameVersion())
-    {
-    case BaseGameVersion::pokeruby:
-        ui->checkBox_NewMap_Allow_Running->setVisible(false);
-        ui->checkBox_NewMap_Allow_Biking->setVisible(false);
-        ui->checkBox_NewMap_Allow_Escape_Rope->setVisible(false);
-        ui->label_NewMap_Allow_Running->setVisible(false);
-        ui->label_NewMap_Allow_Biking->setVisible(false);
-        ui->label_NewMap_Allow_Escape_Rope->setVisible(false);
-        break;
-    case BaseGameVersion::pokeemerald:
-        ui->checkBox_NewMap_Allow_Running->setVisible(true);
-        ui->checkBox_NewMap_Allow_Biking->setVisible(true);
-        ui->checkBox_NewMap_Allow_Escape_Rope->setVisible(true);
-        ui->label_NewMap_Allow_Running->setVisible(true);
-        ui->label_NewMap_Allow_Biking->setVisible(true);
-        ui->label_NewMap_Allow_Escape_Rope->setVisible(true);
-        break;
-    case BaseGameVersion::pokefirered:
-        ui->checkBox_NewMap_Allow_Running->setVisible(true);
-        ui->checkBox_NewMap_Allow_Biking->setVisible(true);
-        ui->checkBox_NewMap_Allow_Escape_Rope->setVisible(true);
-        ui->label_NewMap_Allow_Running->setVisible(true);
-        ui->label_NewMap_Allow_Biking->setVisible(true);
-        ui->label_NewMap_Allow_Escape_Rope->setVisible(true);
-        break;
-    }
-    if (projectConfig.getUseCustomBorderSize()) {
-        if (importedMap) {
-            ui->spinBox_NewMap_BorderWidth->setValue(mapLayout->border_width);
-            ui->spinBox_NewMap_BorderHeight->setValue(mapLayout->border_height);
-        }
-        ui->spinBox_NewMap_BorderWidth->setVisible(true);
-        ui->spinBox_NewMap_BorderHeight->setVisible(true);
-        ui->label_NewMap_BorderWidth->setVisible(true);
-        ui->label_NewMap_BorderHeight->setVisible(true);
-    } else {
-        if (importedMap) {
-            ui->spinBox_NewMap_BorderWidth->setValue(DEFAULT_BORDER_WIDTH);
-            ui->spinBox_NewMap_BorderHeight->setValue(DEFAULT_BORDER_HEIGHT);
-        }
-        ui->spinBox_NewMap_BorderWidth->setVisible(false);
-        ui->spinBox_NewMap_BorderHeight->setVisible(false);
-        ui->label_NewMap_BorderWidth->setVisible(false);
-        ui->label_NewMap_BorderHeight->setVisible(false);
-    }
-    if (projectConfig.getFloorNumberEnabled()) {
-        ui->spinBox_NewMap_Floor_Number->setVisible(true);
-        ui->label_NewMap_Floor_Number->setVisible(true);
-    } else {
-        ui->spinBox_NewMap_Floor_Number->setVisible(false);
-        ui->label_NewMap_Floor_Number->setVisible(false);
-    }
+    // Dimensions and tilesets can't be changed for new maps using an existing layout
+    ui->spinBox_NewMap_Width->setDisabled(true);
+    ui->spinBox_NewMap_Height->setDisabled(true);
+    ui->spinBox_NewMap_BorderWidth->setDisabled(true);
+    ui->spinBox_NewMap_BorderHeight->setDisabled(true);
+    ui->comboBox_NewMap_Primary_Tileset->setDisabled(true);
+    ui->comboBox_NewMap_Secondary_Tileset->setDisabled(true);
 }
 
 void NewMapPopup::on_lineEdit_NewMap_Name_textChanged(const QString &text) {
@@ -264,7 +257,7 @@ void NewMapPopup::on_pushButton_NewMap_Accept_clicked() {
     newMap->name = newMapName;
     newMap->type = this->ui->comboBox_NewMap_Type->currentText();
     newMap->location = this->ui->comboBox_NewMap_Location->currentText();
-    newMap->song = this->ui->comboBox_Song->currentText();
+    newMap->song = this->ui->comboBox_NewMap_Song->currentText();
     newMap->requiresFlash = false;
     newMap->weather = this->project->weatherNames.value(0, "WEATHER_NONE");
     newMap->show_location = this->ui->checkBox_NewMap_Show_Location->isChecked();